# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/teachers": {
            "get": {
                "description": "list teacher entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List teacher entities",
                "operationId": "list-teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Teacher"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create teacher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create teacher",
                "operationId": "create-teacher",
                "parameters": [
                    {
                        "description": "Teacher entity",
                        "name": "teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/teachers/{id}": {
            "get": {
                "description": "get teacher by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a teacher entity by ID",
                "operationId": "get-teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update teacher by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a teacher entity by ID",
                "operationId": "update-teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Teacher entity",
                        "name": "teacher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Teacher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get teacher by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a teacher entity by ID",
                "operationId": "delete-teacher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Teacher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ent.Lessonplan": {
            "type": "object",
            "properties": {
                "Room": {
                    "description": "Room holds the value of the \"Room\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the LessonplanQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LessonplanEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.LessonplanEdges": {
            "type": "object",
            "properties": {
                "courseID": {
                    "description": "CourseID holds the value of the Course_ID edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Subject"
                },
                "groupID": {
                    "description": "GroupID holds the value of the Group_id edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Section"
                },
                "professorID": {
                    "description": "ProfessorID holds the value of the Professor_ID edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Teacher"
                }
            }
        },
        "ent.Section": {
            "type": "object",
            "properties": {
                "Date_Time": {
                    "description": "DateTime holds the value of the \"Date_Time\" field.",
                    "type": "string"
                },
                "Group": {
                    "description": "Group holds the value of the \"Group\" field.",
                    "type": "integer"
                },
                "Room": {
                    "description": "Room holds the value of the \"Room\" field.",
                    "type": "string"
                },
                "Seat_left": {
                    "description": "SeatLeft holds the value of the \"Seat_left\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SectionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SectionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "recieve": {
                    "description": "Recieve holds the value of the \"recieve\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.SectionEdges": {
            "type": "object",
            "properties": {
                "secID": {
                    "description": "SecID holds the value of the sec_id edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Lessonplan"
                    }
                }
            }
        },
        "ent.Subject": {
            "type": "object",
            "properties": {
                "Credit": {
                    "description": "Credit holds the value of the \"Credit\" field.",
                    "type": "integer"
                },
                "Price": {
                    "description": "Price holds the value of the \"Price\" field.",
                    "type": "number"
                },
                "Subject_Name": {
                    "description": "SubjectName holds the value of the \"Subject_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SubjectQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SubjectEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SubjectEdges": {
            "type": "object",
            "properties": {
                "professorID": {
                    "description": "ProfessorID holds the value of the Professor_ID edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Teacher"
                },
                "subjectID": {
                    "description": "SubjectID holds the value of the Subject_ID edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Lessonplan"
                    }
                }
            }
        },
        "ent.Teacher": {
            "type": "object",
            "properties": {
                "Teacher_Name": {
                    "description": "TeacherName holds the value of the \"Teacher_Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TeacherQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TeacherEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TeacherEdges": {
            "type": "object",
            "properties": {
                "teacherID": {
                    "description": "TeacherID holds the value of the Teacher_ID edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Subject"
                    }
                },
                "teachersID": {
                    "description": "TeachersID holds the value of the Teachers_ID edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Lessonplan"
                    }
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}